requirements.txt
cython
python-telegram-bot
pyinstaller